service page/ images /dotView

+++ app/src/main/java/com/blackmirror/dongda/Tools/AYScreenSingleton.java
@@ -1,7 +1,6 @@
 package com.blackmirror.dongda.Tools;
 
 import android.content.Context;
-import android.graphics.Bitmap;
 import android.util.DisplayMetrics;
 import android.view.WindowManager;
 
@@ -47,13 +46,13 @@ public class AYScreenSingleton extends Object {
 
             DisplayMetrics dm = new DisplayMetrics();
             wm.getDefaultDisplay().getMetrics(dm);
-            int width = dm.widthPixels;         // 屏幕宽度（像素）
-            int height = dm.heightPixels;       // 屏幕高度（像素）
+            float width = dm.widthPixels;         // 屏幕宽度（像素）
+            float height = dm.heightPixels;       // 屏幕高度（像素）
             float density = dm.density;         // 屏幕密度（0.75 / 1.0 / 1.5）
 
             // 屏幕宽度:屏幕宽度（像素）/屏幕密度
-            screenWidth = (int) (width / density);  // 屏幕宽度(dp)
-            screenHeight = (int) (height / density);// 屏幕高度(dp)
+            screenWidth = width / density;  // 屏幕宽度(dp)
+            screenHeight = height / density;// 屏幕高度(dp)
             screenDensity = density;
         }
     }
